Changed TagUpdate- to MobDamage- and MobDeath-Listener for further event actions.
